Move getRage and calculateResourceIncrease into utils directory
