+++ src/analysis/retail/warrior/shared/utils/calculateResourceIncrease.ts
